TECH-2322 expose es6 modules: reorg to only have legacy items in legacy/ dirs
