+++ spec/data_table/server_side_data_table_spec.js
@@ -1,4 +1,4 @@
-import { default as Backdraft } from "../../src/legacy/entry";
+import { default as Backdraft } from "../../src/entry";
 
 describe("DataTable Plugin", function() {
   var app;
