+++ spec/listing/listing_spec.js
@@ -3,7 +3,7 @@ import Item from "../../src/listing/item";
 
 import Collection from "../../src/collection";
 
-import { default as Backdraft } from "../../src/legacy/entry";
+import { default as Backdraft } from "../../src/entry";
 
 class Sub extends Item {
   get tagName() { return "li"; }
