+++ src/entry.js
