+++ src/plugin.js
